requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
webdriver-manager>=4.0.0
openai>=2.0.0
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
Pillow>=10.0.0
PyPDF2>=3.0.0
openpyxl>=3.1.0
lxml>=4.9.0
pydantic>=2.0.0
gunicorn>=21.0.0
